기존 플러그인과의 호환성 문제 해결
