+++ source/Grabacr07.KanColleWrapper/KanColleWrapper.csproj
@@ -155,7 +155,6 @@
     <Compile Include="Models\Raw\kcsapi_remodel_slotlist.cs" />
     <Compile Include="Models\Raw\kcsapi_remodel_slotlist_detail.cs" />
     <Compile Include="Models\Raw\kcsapi_slot_exchange_index.cs" />
-    <Compile Include="Models\Raw\map_start_next.cs" />
     <Compile Include="Models\ShipSituation.cs" />
     <Compile Include="Models\ShipSlot.cs" />
     <Compile Include="Models\SlotItemEquipType.cs" />
